+++ src/Form/ResultatsAddType.php
@@ -6,6 +6,7 @@ use App\Entity\Resultats;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\IntegerType;
 use Symfony\Component\Form\Extension\Core\Type\NumberType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Validator\Constraints\NotBlank;
@@ -108,14 +109,24 @@ class ResultatsAddType extends AbstractType
                         'max' => 5,
                         'maxMessage' => 'Le nombre ne peut pas dépasser 5 caractéres',
                     ])
+                ], 
+            ])
+            ->add('commentaire', TextareaType::class, [
+                'attr' => [
+                    'class' => 'form-control',
+                    'placeholder' => 'Ce champ est facultatif',
+                    'rows' => '1',
                 ],
-            ]);
+                'label' => false,
+                'required' => false,
+            ]);     
     }
 
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
             'data_class' => Resultats::class,
+            'sanitize_html' => true,
         ]);
     }
 }