+++ src/Form/ResultatsManageType.php
@@ -8,6 +8,7 @@ use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\IntegerType;
 use Symfony\Component\Form\Extension\Core\Type\NumberType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Validator\Constraints\NotBlank;
@@ -133,6 +134,16 @@ class ResultatsManageType extends AbstractType
                         'maxMessage' => 'Le nombre ne peut pas dépasser 5 caractéres',
                     ])
                 ],
+            ])
+            ->add('commentaire', TextareaType::class, [
+                'attr' => [
+                    'class' => 'form-control',
+                    'placeholder' => 'Ce champ est facultatif',
+                    'value' => $repo->getCommentaire(),
+                    'rows' => '1',
+                ],
+                'label' => false,
+                'required' => false,
             ]);
     }
 
@@ -140,6 +151,7 @@ class ResultatsManageType extends AbstractType
     {
         $resolver->setDefaults([
             'data_class' => Resultats::class,
+            'sanitize_html' => true,
         ]);
     }
 }