+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.2.0
+
+[+] Ajout de l'historique détaillé
+[+] Améliorations mineures
+
 ## 2.1.0
 
 [+] Ajout de la fonctionnalité mot de passe oublié